// src/domain/graph/model/GroupNode.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Annotation } from "./Annotation";

export type GroupNode = { id: string, label: string, parentId: string | null, color?: string, facadeModuleIds: Array<string>, disconnectedByDefault: boolean, disconnectedModuleIds: Array<string>, annotation?: Annotation, architectureDoc?: string, };